Refactored Emulator to allow single opcode steps

+++ emulator/GUITickEventArgs.cs
@@ -7,7 +7,7 @@ namespace JustinCredible.c8emu
     class GUITickEventArgs : EventArgs
     {
         // Out
-        public long ElapsedMilliseconds { get; set; }
+        public double ElapsedMilliseconds { get; set; }
         // public ??? Keys { get; set; } // TODO
 
         // In
